binding/nonnative-ffi/pypolycall/backup_20250603_233522/requirements.txt
# PyPolyCall Core Dependencies - Protocol Compliance
# LibPolyCall Trial v1 Python Binding

# Network and Protocol
aiohttp>=3.8.0
websockets>=11.0.0
httpx>=0.24.0

# Serialization and Data Handling
msgpack>=1.0.0
protobuf>=4.23.0
pydantic>=2.0.0

# Cryptography and Security (Zero-Trust)
cryptography>=41.0.0
PyJWT>=2.8.0
bcrypt>=4.0.0

# Configuration Management
PyYAML>=6.0.0
toml>=0.10.2
python-dotenv>=1.0.0

# State Management and Telemetry
structlog>=23.0.0
attrs>=23.0.0

# FFI and System Integration
cffi>=1.15.0

# Development and Testing (optional)
# pytest>=7.0.0
# pytest-asyncio>=0.21.0
# pytest-cov>=4.0.0
# black>=23.0.0
# mypy>=1.0.0
